+++ tests/opentelemetry-docker-tests/tests/sqlalchemy_tests/test_mysql.py
@@ -68,8 +68,9 @@ class MysqlConnectorTestCase(SQLAlchemyTestMixin):
                 conn.execute("SELECT * FROM a_wrong_table").fetchall()
 
         spans = self.memory_exporter.get_finished_spans()
-        self.assertEqual(len(spans), 1)
-        span = spans[0]
+        # one span for the connection and one for the query
+        self.assertEqual(len(spans), 2)
+        span = spans[1]
         # span fields
         self.assertEqual(span.name, "SELECT opentelemetry-tests")
         self.assertEqual(